🐛fix: #92 스켈레톤 컴포넌트 세부 수정 및 코드 분할
